refactor: restructures qcore instrument directory to allow for individual components to be used without importing all submodules
