--- qcore/instruments/__init__.py
@@ -1,15 +1,0 @@
-""" """
-
-from qcore.instruments.drivers import GS200, LMS, MS46522B, QM, SA124, SC5503B, SC5511A
-from qcore.instruments.instrument import DummyInstrument
-
-__all__ = [
-    "GS200",
-    "LMS",
-    "MS46522B",
-    "QM",
-    "SA124",
-    "SC5503B",
-    "SC5511A",
-    "DummyInstrument",
-]
