+++ qcore/instruments/controllers/__init__.py
@@ -1,0 +1,1 @@
+from qcore.instruments.controllers import QM
