+++ qcore/instruments/controllers/qm.py
@@ -9,9 +9,9 @@ from qm.qua._dsl import _ProgramScope
 
 from qcore.helpers.logger import logger
 from qcore.instruments.instrument import Instrument, ConnectionError
-from qcore.instruments.drivers.qm_config_builder import QMConfigBuilder, QMConfig
-from qcore.instruments.drivers.qm_result_fetcher import QMResultFetcher
-from qcore.instruments.drivers.vaunix_lms import LMS
+from qcore.instruments.controllers.qm_config_builder import QMConfigBuilder, QMConfig
+from qcore.instruments.controllers.qm_result_fetcher import QMResultFetcher
+from qcore.instruments.peripherals.vaunix_lms import LMS
 from qcore.modes.mode import Mode
 
 
@@ -99,4 +99,4 @@ class QM(Instrument):
 
     def set_output_dc_offset_by_element(self, element: str, input: str, offset: float):
         """ """
-        self._qm.set_output_dc_offset_by_element(element, input, offset) 
+        self._qm.set_output_dc_offset_by_element(element, input, offset)
