--- qcore/instruments/drivers/__init__.py
@@ -1,7 +1,0 @@
-from qcore.instruments.drivers.anritsu_ms46522b import MS46522B
-from qcore.instruments.drivers.qm import QM
-from qcore.instruments.drivers.signalcore_sc5503b import SC5503B
-from qcore.instruments.drivers.signalcore_sc5511a import SC5511A
-from qcore.instruments.drivers.signalhound_sa124 import SA124
-from qcore.instruments.drivers.vaunix_lms import LMS
-from qcore.instruments.drivers.yokogawa_gs200 import GS200
