+++ qcore/instruments/peripherals/__init__.py
@@ -1,0 +1,6 @@
+from qcore.instruments.peripherals.anritsu_ms46522b import MS46522B
+from qcore.instruments.peripherals.signalcore_sc5503b import SC5503B
+from qcore.instruments.peripherals.signalcore_sc5511a import SC5511A
+from qcore.instruments.peripherals.signalhound_sa124 import SA124
+from qcore.instruments.peripherals.vaunix_lms import LMS
+from qcore.instruments.peripherals.yokogawa_gs200 import GS200
