+++ qcore/instruments/peripherals/signalcore_sc5503b.dll
