+++ qcore/instruments/peripherals/signalcore_sc5511a.dll
